% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp.R
\name{rejection_sampling_Rcpp}
\alias{rejection_sampling_Rcpp}
\title{Rejection Sampling from Truncated Normal Distribution (Rcpp version)}
\usage{
rejection_sampling_Rcpp(n_samples, mu = 0, sigma = 1, lower = -2, upper = 2)
}
\arguments{
\item{n_samples}{Number of accepted samples desired}

\item{mu}{Mean of the normal distribution. Default is 0.}

\item{sigma}{Standard deviation of the normal distribution. Default is 1.}

\item{lower}{Lower truncation bound. Default is -2.}

\item{upper}{Upper truncation bound. Default is 2.}
}
\value{
A numeric vector of accepted samples from the truncated normal distribution
}
\description{
Generates samples from a truncated normal distribution using rejection sampling.
  This Rcpp implementation demonstrates the inefficiency of push_back operations in Rcpp
  due to vector copying on each growth operation.
}
